+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `is_one_of()` to `property_filter` for compact building of an allow-list based property filter.
 - Add "ABDI1" and "ABDI2" to "extra-indices-dict.json" ([#875](https://github.com/Open-EO/openeo-python-client/pull/875))
+- Support for collection "queryables": more correct warnings when using property filtering in `load_collection` ([#483](https://github.com/Open-EO/openeo-python-client/issues/483))
 
 ### Changed